# swagger_full.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Flash API",
    "version": "1.0.0",
    "description": "API completa para gerenciamento de múltiplas sessões WhatsApp",
    "contact": {
      "name": "API Support",
      "email": "quickgestor@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8087/api",
      "description": "Servidor de Beta"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "apikey",
        "description": "API Key para autenticação"
      }
    },
...
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/config/config": {
      "put": {
        "summary": "Atualizar configurações da sessão",
        "tags": [
          "Configurações"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ignoreGroups": {
                    "type": "boolean",
                    "description": "Ignorar mensagens de grupos",
                    "default": false
                  },
                  "autoRead": {
                    "type": "boolean",
                    "description": "Marcar mensagens como lidas automaticamente",
                    "default": false
                  },
                  "rejectCalls": {
                    "type": "boolean",
                    "description": "Rejeitar chamadas automaticamente",
                    "default": true
                  },
                  "msg_rejectcalls": {
                    "type": "string",
                    "description": "Resposta De regeitar msg",
                    "default": "Não atendemos ligações"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Configurações atualizadas com sucesso"
          }
        }
      }
    },
    "/config/session": {
      "get": {
        "summary": "Obter configurações da sessão",
        "tags": [
          "Configurações"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Configurações da sessão"
          }
        }
      }
    },
    "/config/webhook": {
      "put": {
        "summary": "Configurar webhook da sessão",
        "tags": [
          "Configurações"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "webhookUrl",
                  "events"
                ],
                "properties": {
                  "webhookUrl": {
                    "type": "string",
                    "description": "URL do webhook"
                  },
                  "status_webhook": {
                    "type": "boolean",
                    "description": "true ou false",
                    "default": true
                  },
                  "events": {
                    "type": "array",
                    "description": "Configurar webhook",
                    "default": [
                      "connection_update",
                      "qr_updated",
                      "message_received",
                      "message_update",
                      "chats_set",
                      "chats_update",
                      "contacts_set",
                      "contacts_update",
                      "groups_update",
                      "group_participants_update",
                      "presence_update",
                      "call",
                      "messaging_history_set"
                    ],
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook adicionado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Webhook adicionado com sucesso"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/config/stats": {
      "get": {
        "summary": "Obter estatísticas da sessão",
        "tags": [
          "Configurações"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Estatísticas da sessão"
          }
        }
      }
    },
    "/contact/list": {
      "get": {
        "summary": "Listar contatos da sessão",
        "tags": [
          "Contatos"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de contatos"
          }
        }
      }
    },
    "/contact/sync": {
      "post": {
        "summary": "Sincronizar contatos manualmente",
        "tags": [
          "Contatos"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Contatos sincronizados com sucesso"
          }
        }
      }
    },
    "/contact/force-sync-all": {
      "post": {
        "summary": "Forçar sincronização completa (contatos, chats, grupos)",
        "tags": [
          "Contatos"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sincronização completa iniciada"
          }
        }
      }
    },
    "/contact/profile": {
      "post": {
        "summary": "Obter perfil de um contato",
        "tags": [
          "Contatos"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "jid"
                ],
                "properties": {
                  "jid": {
                    "type": "string",
                    "description": "JID do contato"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Perfil do contato"
          }
        }
      }
    },
    "/contact/check": {
      "post": {
        "summary": "Verificar se números estão no WhatsApp",
        "tags": [
          "Contatos"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "numbers"
                ],
                "properties": {
                  "numbers": {
                    "type": "array",
                    "description": "Lista de números para verificar",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resultado da verificação"
          }
        }
      }
    },
    "/contact/block": {
      "post": {
        "summary": "Bloquear contato",
        "tags": [
          "Contatos"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "jid"
                ],
                "properties": {
                  "jid": {
                    "type": "string",
                    "description": "JID do contato para bloquear"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contato bloqueado com sucesso"
          }
        }
      }
    },
    "/contact/unblock": {
      "post": {
        "summary": "Desbloquear contato",
        "tags": [
          "Contatos"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "jid"
                ],
                "properties": {
                  "jid": {
                    "type": "string",
                    "description": "JID do contato para desbloquear"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contato desbloqueado com sucesso"
          }
        }
      }
    },
...
  }
}